+++ server/src/controllers/tasks/index.ts
@@ -1,8 +1,8 @@
 import addTask from './addTask';
 import getTasks from './getTasks';
 import getTasksSection from './getTasksSection';
-import editTask from './editTaskController';
-import deleteTask from './deleteTaskController';
+import editTask from './editTask';
+import deleteTask from './deleteTask';
 
 export {
   addTask,
