+++ server/src/routes/tasksRoute.ts
@@ -1,12 +1,18 @@
 import express from 'express';
 import { checkAuth } from '../middlewares';
-import { addTask, getTasks, getTasksSection,editTask, deleteTask } from '../controllers';
+import {
+  addTask,
+  getTasks,
+  getTasksSection,
+  editTask,
+  deleteTask,
+} from '../controllers';
 
 const tasksRoute = express.Router();
+tasksRoute.use(checkAuth);
 tasksRoute.post('/addTask', addTask);
-tasksRoute.get('/tasks', checkAuth, getTasks);
+tasksRoute.get('/tasks', getTasks);
 tasksRoute.get('/tasks/:id', getTasksSection);
-tasksRoute.put('/task/:id', checkAuth, editTask);
-tasksRoute.delete('/deleteTask/:id', checkAuth, deleteTask);
-
+tasksRoute.put('/task/:id', editTask);
+tasksRoute.delete('/deleteTask/:id', deleteTask);
 export default tasksRoute;
